+++ designer/server/src/test/scala/pl/touk/nussknacker/ui/api/DeploymentApiHttpServiceBusinessSpec.scala
@@ -1,31 +1,21 @@
 package pl.touk.nussknacker.ui.api
 
-import com.nimbusds.jose.util.StandardCharset
 import com.typesafe.scalalogging.LazyLogging
 import io.restassured.RestAssured.`given`
 import io.restassured.module.scala.RestAssuredSupport.AddThenToResponse
-import org.apache.commons.io.FileUtils
 import org.scalatest.freespec.AnyFreeSpecLike
 import org.scalatest.matchers.should.Matchers
-import org.testcontainers.containers.BindMode
 import pl.touk.nussknacker.engine.build.ScenarioBuilder
-import pl.touk.nussknacker.engine.flink.test.docker.FileSystemBind
 import pl.touk.nussknacker.engine.graph.expression.Expression
 import pl.touk.nussknacker.test.base.it.{NuItTest, WithBatchConfigScenarioHelper}
-import pl.touk.nussknacker.test.config.{
-  WithBatchDesignerConfig,
-  WithBusinessCaseRestAssuredUsersExtensions,
-  WithFlinkContainersDeploymentManager
-}
+import pl.touk.nussknacker.test.config.{WithBatchDesignerConfig, WithBusinessCaseRestAssuredUsersExtensions}
 import pl.touk.nussknacker.test.{NuRestAssureMatchers, RestAssuredVerboseLogging, VeryPatientScalaFutures}
-
-import java.nio.file.Files
 
 class DeploymentApiHttpServiceBusinessSpec
     extends AnyFreeSpecLike
     with NuItTest
     with WithBatchDesignerConfig
-    with WithFlinkContainersDeploymentManager
+    with BaseDeploymentApiHttpServiceBusinessSpec
     with WithBatchConfigScenarioHelper
     with WithBusinessCaseRestAssuredUsersExtensions
     with NuRestAssureMatchers
@@ -34,50 +24,13 @@ class DeploymentApiHttpServiceBusinessSpec
     with VeryPatientScalaFutures
     with Matchers {
 
-  private lazy val outputDirectory =
-    Files.createTempDirectory(s"nusssknacker-${getClass.getSimpleName}-transactions_summary-")
-
-  private lazy val tablesDefinitionBind = FileSystemBind(
-    "designer/server/src/test/resources/config/business-cases/tables-definition.sql",
-    "/opt/flink/designer/server/src/test/resources/config/business-cases/tables-definition.sql",
-    BindMode.READ_ONLY
-  )
-
-  private lazy val inputTransactionsBind = FileSystemBind(
-    "designer/server/src/test/resources/transactions",
-    "/transactions",
-    BindMode.READ_ONLY
-  )
-
-  private lazy val outputTransactionsSummaryBind = FileSystemBind(
-    outputDirectory.toString,
-    "/output/transactions_summary",
-    BindMode.READ_WRITE
-  )
-
-  override protected def jobManagerExtraFSBinds: List[FileSystemBind] =
-    List(
-      tablesDefinitionBind,
-      // input must be also available on the JM side to allow their to split work into multiple subtasks
-      inputTransactionsBind
-    )
-
-  override protected def taskManagerExtraFSBinds: List[FileSystemBind] = {
-
-    List(
-      // table definitions must be also on the TM side. This is necessary because we create full model definition for the purpose of
-      // interpreter used in scenario parts using built-in components - TODO: it shouldn't be needed
-      tablesDefinitionBind,
-      inputTransactionsBind,
-      outputTransactionsSummaryBind
-    )
-  }
-
   private val scenarioName = "batch-test"
+
+  private val sourceNodeId = "fooSourceNodeId"
 
   private val scenario = ScenarioBuilder
     .streaming(scenarioName)
-    .source("source", "table", "Table" -> Expression.spel("'transactions'"))
+    .source(sourceNodeId, "table", "Table" -> Expression.spel("'transactions'"))
     .emptySink(
       "sink",
       "table",
@@ -85,12 +38,13 @@ class DeploymentApiHttpServiceBusinessSpec
       "Value" -> Expression.spel("#input")
     )
 
-  override protected def afterAll(): Unit = {
-    FileUtils.deleteQuietly(outputDirectory.toFile) // it might not work because docker user can has other uid
-    super.afterAll()
-  }
+  private val correctDeploymentRequest = s"""{
+                                            |  "nodesDeploymentData": {
+                                            |    "$sourceNodeId": "`date` = '2024-01-01'"
+                                            |  }
+                                            |}""".stripMargin
 
-  "The endpoint for deployment requesting" - {
+  "The deployment requesting endpoint" - {
     "authenticated as user with deploy access should" - {
       "run deployment" in {
         val requestedDeploymentId = "some-requested-deployment-id"
@@ -100,7 +54,7 @@ class DeploymentApiHttpServiceBusinessSpec
           }
           .when()
           .basicAuthAdmin()
-          .jsonBody("{}")
+          .jsonBody(correctDeploymentRequest)
           .put(s"$nuDesignerHttpAddress/api/scenarios/$scenarioName/deployments/$requestedDeploymentId")
           .Then()
           // TODO (next PRs): we should return 201 and we should check status of deployment before we verify output
@@ -118,7 +72,7 @@ class DeploymentApiHttpServiceBusinessSpec
             createSavedScenario(scenario)
           }
           .when()
-          .jsonBody("{}")
+          .jsonBody(correctDeploymentRequest)
           .put(s"$nuDesignerHttpAddress/api/scenarios/$scenarioName/deployments/foo-deployment-id")
           .Then()
           .statusCode(401)
@@ -133,7 +87,7 @@ class DeploymentApiHttpServiceBusinessSpec
           }
           .when()
           .basicAuthUnknownUser()
-          .jsonBody("{}")
+          .jsonBody(correctDeploymentRequest)
           .put(s"$nuDesignerHttpAddress/api/scenarios/$scenarioName/deployments/foo-deployment-id")
           .Then()
           .statusCode(401)
@@ -148,7 +102,7 @@ class DeploymentApiHttpServiceBusinessSpec
           }
           .when()
           .basicAuthNoPermUser()
-          .jsonBody("{}")
+          .jsonBody(correctDeploymentRequest)
           .put(s"$nuDesignerHttpAddress/api/scenarios/$scenarioName/deployments/foo-deployment-id")
           .Then()
           .statusCode(403)
@@ -163,31 +117,12 @@ class DeploymentApiHttpServiceBusinessSpec
           }
           .when()
           .basicAuthWriter()
-          .jsonBody("{}")
+          .jsonBody(correctDeploymentRequest)
           .put(s"$nuDesignerHttpAddress/api/scenarios/$scenarioName/deployments/foo-deployment-id")
           .Then()
           .statusCode(403)
       }
     }
-  }
-
-  private def outputTransactionSummaryContainsResult(): Unit = {
-
-    // finished deploy doesn't mean that processing is finished
-    // TODO (next PRs): we need to wait for the job completed status instead
-    val transactionsSummaryFiles = eventually {
-      val files = Option(outputDirectory.toFile.listFiles(!_.isHidden)).toList.flatten
-      files should have size 1
-      files
-    }
-    val transactionsSummaryContent =
-      FileUtils.readFileToString(transactionsSummaryFiles.head, StandardCharset.UTF_8)
-    // TODO (next PRs): aggregate by clientId
-    transactionsSummaryContent should include(
-      """client1,1.12
-        |client2,2.21
-        |client1,3""".stripMargin
-    )
   }
 
 }